requirements.txt
aiohttp
async_timeout
chardet
discord.py
multidict
requests
websockets
